django/requirements.txt
Django==4.2.0
django-cors-headers>=4.2.0
djangorestframework>=3.11,<4.0
psycopg2-binary>=2.8
redis>=4.0
requests>=2.31.0
celery==5.2.7
django-celery-beat==2.5.0
django-celery-results==2.3.0
boto3==1.24.0
djangorestframework-simplejwt==5.2.2